requirements.txt
discord.py~=2.5.2
psutil~=7.0.0
python-dotenv~=1.1.0